requirements.txt
opencv-python
flask
ultralytics
Pillow
python-dotenv
google-generativeai
